backend/imd_dev_web/requirements.txt
asgiref==3.8.1
Django==5.1.1
sqlparse==0.5.1
psycopg2==2.9.9
Pillow>=9.0.0
djangorestframework==3.14.0
drf-yasg==1.21.7
requests==2.28.1